+++ interceptors/subscription-event-interceptors/src/test/java/uk/gov/justice/services/event/source/subscriptions/interceptors/it/SubscriptionEventInterceptorIT.java
@@ -29,6 +29,7 @@ import uk.gov.justice.services.core.annotation.Adapter;
 import uk.gov.justice.services.core.annotation.Handles;
 import uk.gov.justice.services.core.annotation.ServiceComponent;
 import uk.gov.justice.services.core.dispatcher.DispatcherCache;
+import uk.gov.justice.services.core.dispatcher.DispatcherConfiguration;
 import uk.gov.justice.services.core.dispatcher.DispatcherFactory;
 import uk.gov.justice.services.core.dispatcher.EmptySystemUserProvider;
 import uk.gov.justice.services.core.dispatcher.EnvelopePayloadTypeConverter;
@@ -216,7 +217,9 @@ public class SubscriptionEventInterceptorIT {
             InitialContextProducer.class,
 
             SchemaValidationErrorMessageGenerator.class,
-            EventSourceNameCalculator.class
+            EventSourceNameCalculator.class,
+
+            DispatcherConfiguration.class
     })
 
     public WebApp war() {